apply new method to calculate base height for cesium camera placement from frustum pixelsize and distance
