+++ libraries/mapping/engines/cesium/src/lib/utils/leafletHelpers.ts
@@ -1,0 +1,14 @@
+// TODO MOVE TO COMMON HELPER LIBRARY IF TESTED AND NEEDED ELSEWHERE
+
+export const isLeafletZoomValid = (zoom: number) => {
+  if (
+    zoom === undefined ||
+    Number.isNaN(zoom) ||
+    zoom === Infinity ||
+    zoom === -Infinity
+  ) {
+    return false;
+  } else {
+    return true;
+  }
+};
